Add commands for  X Y A B buttons on ROKStar controller
Adds commands for the buttons that are on the ROKStar controller.
XY_MOTOR_CLOCKWISE isn't immediately obvious which button on the controller that is associated with. It is good if a command is sent from ROKduino to a MRB, but not necessarily as good if sending commands from the controller to the ROKduino. All the ROKduino sets come with the controller now, so it makes sense to have easily readable commands for the buttons, at least for the labeled buttons. The directional pad is fairly obvious that forward, backward, etc correspond with the Up and Down buttons.